refactor: split large dashboard, todos, and completions components (Phase 5)
- Dashboard route: extract useDashboardTileSaver (debounced tile saves,
  optimistic cache write, legacy-id self-heal) and the LayoutManagerDialogs
  trio; prune the -components barrel re-exports the route no longer uses
- TodosEditor: extract TodoReadRow and hoist the display comparator to a
  named module-level function
- DailyCompletionsManager: extract CompletionsMonthNav and
  DailyCompletionEntryRow, retiring the fallow complexity suppression on
  the old inline row callback
