+++ packages/client/src/routes/dashboard/-components/layout/index.ts
@@ -1,4 +1,6 @@
 export { AddLayoutDialog } from "./-AddLayoutDialog";
 export { DashboardGrid } from "./-DashboardGrid";
+export { LayoutManagerDialogs } from "./-LayoutManagerDialogs";
 export { LayoutTab } from "./-LayoutTab";
+export { useDashboardTileSaver } from "./-useDashboardTileSaver";
 export { VisibleTilesDialog } from "./-VisibleTilesDialog";
